pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "densigrav"
version = "0.1.0"
description = "Gravity anomaly preprocessing (CBA) and 2D section tools for density-contrast interpretation."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Yoko Koyabu", email = "yoko.koyabu@gmail.com" }
]
keywords = ["geophysics", "gravity", "bouguer", "terrain-correction", "qgis"]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only"
]

dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
  "black>=24.0",
  "build>=1.0",
]
geo = [
  "pandas>=2.0",
  "geopandas>=0.14",
  "pyogrio>=0.7",
  "shapely>=2.0",
  "pyproj>=3.6",
]
raster = [
  "numpy>=1.23",
  "rasterio>=1.3",
]
grav = [
  "harmonica>=0.7.0",
  "xarray>=2023.0",
]
grid = [
  "verde>=1.8.0",
  "scipy>=1.10",
]
viz = [
  "matplotlib>=3.8",
]

[project.scripts]
densigrav = "densigrav.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py39"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py39"]

